Refactor of more code
