+++ irispy/utils/tests/test_moments.py
@@ -18,7 +18,7 @@ def test_calculate_moments_basic(sns_sg_file):
     Test that calculate_moments runs on real data and returns correct shapes and units.
     """
     raster_collection = read_files(sns_sg_file)
-    cube = raster_collection["C II 1336"][0]
+    cube = raster_collection["C II 1336"]
     rest_wvl = 1332.9 * u.Angstrom
     moments = calculate_moments(cube, rest_wavelength=rest_wvl, wings=0.1 * u.Angstrom)
     assert isinstance(moments, RasterCollection)
@@ -58,7 +58,7 @@ def test_calculate_moments_sliced_cube(sns_sg_file):
     Test that calculate_moments works on a sliced cube and with default arguments.
     """
     raster_collection = read_files(sns_sg_file)
-    cube = raster_collection["C II 1336"][0]
+    cube = raster_collection["C II 1336"]
     cube_slice = cube[10, :, :]
     moments = calculate_moments(cube_slice)
     assert set(moments.keys()) == {"intensity", "centroid", "width"}
@@ -72,7 +72,7 @@ def test_calculate_moments_asymmetric_wings(sns_sg_file):
     Test that calculate_moments works with asymmetric wings.
     """
     raster_collection = read_files(sns_sg_file)
-    cube = raster_collection["C II 1336"][0]
+    cube = raster_collection["C II 1336"]
     rest_wvl = 1332.9 * u.Angstrom
     moments = calculate_moments(cube, rest_wavelength=rest_wvl, wings=(0.05, 0.15) * u.Angstrom)
     assert set(moments.keys()) == {"intensity", "centroid", "width", "velocity", "velocity_width"}
@@ -110,7 +110,7 @@ def test_calculate_moments_wings_without_rest_wavelength(sns_sg_file):
     rest_wavelength.
     """
     raster_collection = read_files(sns_sg_file)
-    cube = raster_collection["C II 1336"][0]
+    cube = raster_collection["C II 1336"]
     with pytest.raises(ValueError, match="rest_wavelength must be provided"):
         calculate_moments(cube, wings=1.0 * u.Angstrom)
 
@@ -475,7 +475,7 @@ def test_calculate_moments_preserves_time_without_spectral_global_coord(sns_sg_f
     Test that moment maps keep scan times without adding a fixed wavelength coordinate.
     """
     raster_collection = read_files(sns_sg_file)
-    cube = raster_collection["C II 1336"][0]
+    cube = raster_collection["C II 1336"]
     moments = calculate_moments(cube)
     intensity = moments["intensity"]
     assert "time" in tuple(intensity.extra_coords.keys())
